Split PHP >= 7.4 tests
