+++ phpstan-php-lt-7.4.neon
@@ -2,6 +2,7 @@ parameters:
 	level: max
 	excludes_analyse:
 		- 'tests/SimpleCli/SimpleCliCommand/TypeHint.php'
+		- 'tests/SimpleCli/Traits/DocumentationTypeHintTest.php'
 	paths:
 		- src
 		- tests
